Fixed scale being converted from double to int (#1036)
## Rationale
Scale was being incorrectly parsed to int when it should be a double for
room size.

## Related Pull Requests
- <!-- list of links to related pull requests (replace this comment) -->

## Changes
- Updated java code to be double instead of int for scale

+++ CageUI/src/org/labkey/cageui/action/RoomHistoryForm.java
@@ -22,7 +22,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class RoomHistoryForm
 {
-    private int _scale;
+    private double _scale;
     @JsonProperty("border_width")
     private int _borderWidth;
     @JsonProperty("border_height")
@@ -30,12 +30,12 @@ public class RoomHistoryForm
     @JsonProperty("historyid")
     private String _historyId;
 
-    public int getScale()
+    public double getScale()
     {
         return _scale;
     }
 
-    public void setScale(int scale)
+    public void setScale(double scale)
     {
         _scale = scale;
     }
